Move CorePagePropertyProvider from Domain to Persistence
It decides which properties to project into the graph database,
which is a persistence concern rather than domain logic.
