+++ src/Persistence/CorePagePropertyProvider.php
@@ -2,7 +2,10 @@
 
 declare( strict_types = 1 );
 
-namespace ProfessionalWiki\NeoWiki\Domain\Page;
+namespace ProfessionalWiki\NeoWiki\Persistence;
+
+use ProfessionalWiki\NeoWiki\Domain\Page\PagePropertyProvider;
+use ProfessionalWiki\NeoWiki\Domain\Page\PagePropertyProviderContext;
 
 class CorePagePropertyProvider implements PagePropertyProvider {
